api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
pydantic>=2.6
requests>=2.31
apscheduler>=3.10
anthropic>=0.25
psutil>=5.9
python-multipart>=0.0.9
slowapi>=0.1.9
PyMuPDF>=1.24
